Busqueda corregida

+++ src/modules/core/favoritos.js
@@ -5,6 +5,7 @@ document.addEventListener("DOMContentLoaded", function () {
   // Inicializar el contador al cargar la página
   if (idUsuario) {
     actualizarContador(idUsuario);
+    fetchFavoritos();
   }
 
   document.getElementById("filterBtn").addEventListener("click", function () {
@@ -47,26 +48,7 @@ document.addEventListener("DOMContentLoaded", function () {
     })
       .then((response) => response.json())
       .then((data) => {
-        let filteredProducts = data;
-
-        const precioCheckbox = document.getElementById("precioCheckbox");
-        const nombreCheckbox = document.getElementById("nombreCheckbox");
-        const precioSelect = document.getElementById("precioSelect").value;
-        const nombreSelect = document.getElementById("nombreSelect").value;
-
-        if (precioCheckbox.checked) {
-          filteredProducts.sort((a, b) => {
-            return precioSelect === "asc" ? a.precio - b.precio : b.precio - a.precio;
-          });
-        }
-
-        if (nombreCheckbox.checked) {
-          filteredProducts.sort((a, b) => {
-            return nombreSelect === "asc" ? a.nombre.localeCompare(b.nombre) : b.nombre.localeCompare(a.nombre);
-          });
-        }
-
-        displayProducts(filteredProducts);
+        displayProducts(data); // Mostrar todos los productos inicialmente
       })
       .catch((error) => console.error("Error fetching favoritos:", error));
   }
@@ -75,45 +57,49 @@ document.addEventListener("DOMContentLoaded", function () {
   function displayProducts(products) {
     const favoritosBody = document.getElementById("favoritosBody");
     favoritosBody.innerHTML = ""; // Limpiar el contenido anterior
-    products.forEach(function (producto) {
-      var productoDiv = document.createElement("div");
-      productoDiv.classList.add("producto");
+    if (products.length === 0) {
+      favoritosBody.innerHTML = "<p>No hay productos favoritos.</p>";
+    } else {
+      products.forEach(function (producto) {
+        var productoDiv = document.createElement("div");
+        productoDiv.classList.add("producto");
 
-      var deleteIcon = document.createElement("span");
-      deleteIcon.innerHTML = '<ion-icon name="trash-outline"></ion-icon>';
-      deleteIcon.classList.add("delete-icon");
-      deleteIcon.addEventListener("click", function () {
-        eliminarProducto(producto.id_producto, productoDiv);
+        var deleteIcon = document.createElement("span");
+        deleteIcon.innerHTML = '<ion-icon name="trash-outline"></ion-icon>';
+        deleteIcon.classList.add("delete-icon");
+        deleteIcon.addEventListener("click", function () {
+          eliminarProducto(producto.id_producto, productoDiv);
+        });
+
+        var nombre = document.createElement("h3");
+        nombre.textContent = producto.nombre;
+        nombre.classList.add("nombre");
+
+        var descripcion = document.createElement("p");
+        descripcion.textContent = producto.descripcion;
+        descripcion.classList.add("descripcion");
+
+        var costo = document.createElement("p");
+        costo.textContent = "Costo: " + parseFloat(producto.precio).toFixed(2); // Formatear a un decimal
+        costo.classList.add("precio");
+
+        var imagen = document.createElement("img");
+        imagen.src = producto.imagen;
+        imagen.classList.add("producto-imagen");
+
+        var infoDiv = document.createElement("div");
+        infoDiv.classList.add("producto-info");
+        infoDiv.appendChild(nombre);
+        infoDiv.appendChild(descripcion);
+        infoDiv.appendChild(costo);
+
+        productoDiv.appendChild(deleteIcon);
+        productoDiv.appendChild(imagen);
+        productoDiv.appendChild(infoDiv);
+
+        favoritosBody.appendChild(productoDiv);
       });
-
-      var nombre = document.createElement("h3");
-      nombre.textContent = producto.nombre;
-      nombre.classList.add("nombre");
-
-      var descripcion = document.createElement("p");
-      descripcion.textContent = producto.descripcion;
-      descripcion.classList.add("descripcion");
-
-      var costo = document.createElement("p");
-      costo.textContent = "Costo: " + parseFloat(producto.precio).toFixed(2); // Formatear a un decimal
-      costo.classList.add("precio");
-
-      var imagen = document.createElement("img");
-      imagen.src = producto.imagen;
-      imagen.classList.add("producto-imagen");
-
-      var infoDiv = document.createElement("div");
-      infoDiv.classList.add("producto-info");
-      infoDiv.appendChild(nombre);
-      infoDiv.appendChild(descripcion);
-      infoDiv.appendChild(costo);
-
-      productoDiv.appendChild(deleteIcon);
-      productoDiv.appendChild(imagen);
-      productoDiv.appendChild(infoDiv);
-
-      favoritosBody.appendChild(productoDiv);
-    });
+    }
   }
 
   function eliminarProducto(id_producto, productoDiv) {
@@ -156,8 +142,8 @@ document.addEventListener("DOMContentLoaded", function () {
 
   // Abrir el modal al hacer clic en el botón
   openfavBtn.addEventListener("click", function () {
-    fetchFavoritos();
     fav.style.display = "block";
+    fetchFavoritos();
   });
 
   // Cerrar el modal al hacer clic en el botón de cerrar
@@ -200,21 +186,29 @@ document.addEventListener("DOMContentLoaded", function () {
   function searchProducts() {
     var input = document.getElementById("searchInput");
     var filter = input.value.toLowerCase();
-    var productos = document.getElementById("favoritosBody").getElementsByClassName("producto");
-    var filteredProducts = [];
 
-    for (var i = 0; i < productos.length; i++) {
-      var nombre = productos[i].getElementsByClassName("nombre")[0];
-      if (nombre && nombre.textContent.toLowerCase().includes(filter)) {
-        filteredProducts.push({
-          nombre: nombre.textContent,
-          descripcion: productos[i].getElementsByClassName("descripcion")[0].textContent,
-          precio: productos[i].getElementsByClassName("precio")[0].textContent.split(": ")[1],
-          imagen: productos[i].getElementsByClassName("producto-imagen")[0].src,
-        });
-      }
+    if (filter === "") {
+      fetchFavoritos(); // Si el campo está vacío, mostrar todos los productos
+    } else {
+      // Limpiar la búsqueda anterior y actualizar la lista
+      fetch("http://localhost/GoCan/src/modules/php/favoritos.php", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ id_usuario: localStorage.getItem("id_usuario") }),
+      })
+        .then((response) => response.json())
+        .then((data) => {
+          // Filtrar productos usando la nueva palabra de búsqueda
+          var filteredProducts = data.filter(function (producto) {
+            return producto.nombre.toLowerCase().includes(filter) ||
+                   producto.descripcion.toLowerCase().includes(filter);
+          });
+
+          displayProducts(filteredProducts); // Mostrar productos filtrados
+        })
+        .catch((error) => console.error("Error fetching favoritos:", error));
     }
-
-    displayProducts(filteredProducts);
   }
 });
